+++ harfbuzz/src/hb-font.cc
@@ -534,7 +534,6 @@ static hb_font_t *_hb_font_create(hb_face_t *face)
     hb_face_make_immutable(face);
     font->parent = hb_font_get_empty();
     font->face = hb_face_reference(face);
-    font->data.init0(font);
     font->x_scale = font->y_scale = hb_face_get_upem(face);
     font->x_mult = font->y_mult = 1 << 16;
 
@@ -657,8 +656,6 @@ void hb_font_destroy(hb_font_t *font)
 {
     if (!hb_object_destroy(font))
         return;
-
-    font->data.fini();
 
     hb_font_destroy(font->parent);
     hb_face_destroy(font->face);