+++ harfbuzz/src/hb-ot-shape.hh
@@ -32,25 +32,6 @@
 #include "hb-ot-map.hh"
 #include "hb-aat-map.hh"
 
-struct hb_ot_shape_plan_key_t
-{
-    unsigned int variations_index[2];
-
-    void init(hb_face_t *face, const int *coords, unsigned num_coords)
-    {
-        for (unsigned int table_index = 0; table_index < 2; table_index++)
-            hb_ot_layout_table_find_feature_variations(
-                face, table_tags[table_index], coords, num_coords, &variations_index[table_index]);
-    }
-
-    bool equal(const hb_ot_shape_plan_key_t *other)
-    {
-        return 0 == memcmp(this, other, sizeof(*this));
-    }
-};
-
-struct hb_shape_plan_key_t;
-
 struct hb_ot_shape_plan_t
 {
     hb_segment_properties_t props;
@@ -131,7 +112,10 @@ struct hb_ot_shape_plan_t
         map.collect_lookups(table_index, lookups);
     }
 
-    HB_INTERNAL bool init0(hb_face_t *face, const hb_shape_plan_key_t *key);
+    HB_INTERNAL bool init0(hb_face_t *face,
+                           const hb_segment_properties_t *props,
+                           const hb_feature_t *user_features,
+                           unsigned int num_user_features, unsigned int *variations_index);
     HB_INTERNAL void fini();
 
     HB_INTERNAL void substitute(hb_font_t *font, hb_buffer_t *buffer) const;
@@ -158,7 +142,7 @@ struct hb_ot_shape_planner_t
 
     HB_INTERNAL hb_ot_shape_planner_t(hb_face_t *face, const hb_segment_properties_t *props);
 
-    HB_INTERNAL void compile(hb_ot_shape_plan_t &plan, const hb_ot_shape_plan_key_t &key);
+    HB_INTERNAL void compile(hb_ot_shape_plan_t &plan, unsigned int *variations_index);
 };
 
 #endif /* HB_OT_SHAPE_HH */