+++ harfbuzz/src/hb-shape-plan.h
@@ -33,6 +33,7 @@
 
 #include "hb-common.h"
 #include "hb-font.h"
+#include "hb-buffer.h"
 
 HB_BEGIN_DECLS
 
@@ -42,29 +43,8 @@ HB_EXTERN hb_shape_plan_t *hb_shape_plan_create(hb_face_t *face,
                                                 const hb_segment_properties_t *props,
                                                 const hb_feature_t *user_features,
                                                 unsigned int num_user_features,
-                                                const char *const *shaper_list);
-
-HB_EXTERN hb_shape_plan_t *hb_shape_plan_create_cached(hb_face_t *face,
-                                                       const hb_segment_properties_t *props,
-                                                       const hb_feature_t *user_features,
-                                                       unsigned int num_user_features,
-                                                       const char *const *shaper_list);
-
-HB_EXTERN hb_shape_plan_t *hb_shape_plan_create2(hb_face_t *face,
-                                                 const hb_segment_properties_t *props,
-                                                 const hb_feature_t *user_features,
-                                                 unsigned int num_user_features,
-                                                 const int *coords,
-                                                 unsigned int num_coords,
-                                                 const char *const *shaper_list);
-
-HB_EXTERN hb_shape_plan_t *hb_shape_plan_create_cached2(hb_face_t *face,
-                                                        const hb_segment_properties_t *props,
-                                                        const hb_feature_t *user_features,
-                                                        unsigned int num_user_features,
-                                                        const int *coords,
-                                                        unsigned int num_coords,
-                                                        const char *const *shaper_list);
+                                                const int *coords,
+                                                unsigned int num_coords);
 
 HB_EXTERN hb_shape_plan_t *hb_shape_plan_get_empty(void);
 
@@ -77,8 +57,6 @@ HB_EXTERN hb_bool_t hb_shape_plan_execute(hb_shape_plan_t *shape_plan,
                                           hb_buffer_t *buffer,
                                           const hb_feature_t *features,
                                           unsigned int num_features);
-
-HB_EXTERN const char *hb_shape_plan_get_shaper(hb_shape_plan_t *shape_plan);
 
 HB_END_DECLS