+++ harfbuzz/src/hb-shape.cc
@@ -28,7 +28,6 @@
 
 #include "hb.hh"
 
-#include "hb-shaper.hh"
 #include "hb-shape-plan.hh"
 #include "hb-buffer.hh"
 #include "hb-font.hh"
@@ -46,66 +45,8 @@
  * contains the output glyphs and their positions.
  **/
 
-#if HB_USE_ATEXIT
-static void free_static_shaper_list();
-#endif
-
-static const char *nil_shaper_list[] = {nullptr};
-
-static struct hb_shaper_list_lazy_loader_t : hb_lazy_loader_t<const char *, hb_shaper_list_lazy_loader_t>
-{
-    static const char **create()
-    {
-        const char **shaper_list = (const char **)calloc(1 + HB_SHAPERS_COUNT, sizeof(const char *));
-        if (unlikely(!shaper_list))
-            return nullptr;
-
-        const hb_shaper_entry_t *shapers = _hb_shapers_get();
-        unsigned int i;
-        for (i = 0; i < HB_SHAPERS_COUNT; i++)
-            shaper_list[i] = shapers[i].name;
-        shaper_list[i] = nullptr;
-
-#if HB_USE_ATEXIT
-        atexit(free_static_shaper_list);
-#endif
-
-        return shaper_list;
-    }
-    static void destroy(const char **l)
-    {
-        free(l);
-    }
-    static const char **get_null()
-    {
-        return nil_shaper_list;
-    }
-} static_shaper_list;
-
-#if HB_USE_ATEXIT
-static void free_static_shaper_list()
-{
-    static_shaper_list.free_instance();
-}
-#endif
-
 /**
- * hb_shape_list_shapers:
- *
- * Retrieves the list of shapers supported by HarfBuzz.
- *
- * Return value: (transfer none) (array zero-terminated=1): an array of
- *    constant strings
- *
- * Since: 0.9.2
- **/
-const char **hb_shape_list_shapers()
-{
-    return static_shaper_list.get_unconst();
-}
-
-/**
- * hb_shape_full:
+ * hb_shape:
  * @font: an #hb_font_t to use for shaping
  * @buffer: an #hb_buffer_t to shape
  * @features: (array length=num_features) (allow-none): an array of user
@@ -122,39 +63,14 @@ const char **hb_shape_list_shapers()
  *
  * Since: 0.9.2
  **/
-hb_bool_t hb_shape_full(hb_font_t *font,
-                        hb_buffer_t *buffer,
-                        const hb_feature_t *features,
-                        unsigned int num_features,
-                        const char *const *shaper_list)
+hb_bool_t hb_shape(hb_font_t *font, hb_buffer_t *buffer, const hb_feature_t *features, unsigned int num_features)
 {
-    hb_shape_plan_t *shape_plan = hb_shape_plan_create_cached2(
-        font->face, &buffer->props, features, num_features, font->coords, font->num_coords, shaper_list);
+    hb_shape_plan_t *shape_plan =
+        hb_shape_plan_create(font->face, &buffer->props, features, num_features, font->coords, font->num_coords);
     hb_bool_t res = hb_shape_plan_execute(shape_plan, font, buffer, features, num_features);
     hb_shape_plan_destroy(shape_plan);
 
     if (res)
         buffer->content_type = HB_BUFFER_CONTENT_TYPE_GLYPHS;
     return res;
-}
-
-/**
- * hb_shape:
- * @font: an #hb_font_t to use for shaping
- * @buffer: an #hb_buffer_t to shape
- * @features: (array length=num_features) (allow-none): an array of user
- *    specified #hb_feature_t or %NULL
- * @num_features: the length of @features array
- *
- * Shapes @buffer using @font turning its Unicode characters content to
- * positioned glyphs. If @features is not %NULL, it will be used to control the
- * features applied during shaping. If two @features have the same tag but
- * overlapping ranges the value of the feature with the higher index takes
- * precedence.
- *
- * Since: 0.9.2
- **/
-void hb_shape(hb_font_t *font, hb_buffer_t *buffer, const hb_feature_t *features, unsigned int num_features)
-{
-    hb_shape_full(font, buffer, features, num_features, nullptr);
 }