+++ harfbuzz/src/hb-shape.h
@@ -39,15 +39,10 @@
 
 HB_BEGIN_DECLS
 
-HB_EXTERN void hb_shape(hb_font_t *font, hb_buffer_t *buffer, const hb_feature_t *features, unsigned int num_features);
-
-HB_EXTERN hb_bool_t hb_shape_full(hb_font_t *font,
-                                  hb_buffer_t *buffer,
-                                  const hb_feature_t *features,
-                                  unsigned int num_features,
-                                  const char *const *shaper_list);
-
-HB_EXTERN const char **hb_shape_list_shapers(void);
+HB_EXTERN hb_bool_t hb_shape(hb_font_t *font,
+                             hb_buffer_t *buffer,
+                             const hb_feature_t *features,
+                             unsigned int num_features);
 
 HB_END_DECLS