+++ harfbuzz/src/hb-shaper-impl.hh
@@ -29,7 +29,6 @@
 
 #include "hb.hh"
 
-#include "hb-shaper.hh"
 #include "hb-face.hh"
 #include "hb-font.hh"
 #include "hb-shape-plan.hh"